+++ src/lw_common/ui/format/column_formatter.cs
@@ -22,7 +22,13 @@ namespace lw_common.ui.format {
             // the text from the cell above (if any)
             public readonly string prev_text;
 
-            public format_cell(match_item item, log_view parent, int col_idx, info_type col_type, formatted_text text, int row_index, int top_row_index, string prev_text) {
+            public enum location_type {
+                view, details_pane, msg_description, smart_edit
+            }
+
+            public readonly location_type location;
+
+            public format_cell(match_item item, log_view parent, int col_idx, info_type col_type, formatted_text text, int row_index, int top_row_index, string prev_text, location_type location) {
                 this.item = item;
                 this.parent = parent;
                 this.col_idx = col_idx;
@@ -31,6 +37,7 @@ namespace lw_common.ui.format {
                 this.row_index = row_index;
                 this.top_row_index = top_row_index;
                 this.prev_text = prev_text;
+                this.location = location;
 
                 fg_color = item.fg(parent);
                 bg_color = item.bg(parent);