+++ src/lw_common/ui/log_view/match/match_item.cs
@@ -194,7 +194,7 @@ namespace lw_common.ui {
         }
 
         // returns the overrides, sorted by index in the string to print
-        public formatted_text override_print(log_view parent, string text, int col_idx) {
+        public formatted_text override_print(log_view parent, string text, int col_idx, column_formatter.format_cell.location_type location) {
             int row_idx = parent.item_index(this);
             int top_row_idx = parent.top_row_idx;
             string prev_text = "";
@@ -202,7 +202,7 @@ namespace lw_common.ui {
                 prev_text = log_view_cell.cell_value(parent.item_at(row_idx - 1), col_idx);
 
             var cell = new column_formatter.format_cell(this, parent, col_idx, log_view_cell.cell_idx_to_type(col_idx), new formatted_text(text),
-                row_idx, top_row_idx, prev_text);
+                row_idx, top_row_idx, prev_text, location);
             parent.formatter.format_before(cell);
             var print = override_print_from_all_places(parent, cell.format_text.text, col_idx);
             cell.format_text.add_parts( print);